Sharing is functional on Vista 64 now.
git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@6758 db2431b9-df06-0410-971c-a38ea9a75485

+++ ifolder/src/UI/Windows/iFolderShell/Win32.cs
@@ -353,7 +353,7 @@ namespace Novell.Win32Util
 			/// <summary>
 			/// 
 			/// </summary>
-			public int idFrom;
+			public IntPtr idFrom;
 			/// <summary>
 			/// 
 			/// </summary>
